Specify pem file encoding.
Without the encoding specified, the pem file is read in as a buffer
instead of a string.
This would then fail the argument validation requirement for
credentials.pem since it wanted a string and got a buffer.
Without the encoding I was unable to run this example.
With the encoding specified, I am able to run the example.