Add support for x509 pem files using `BEGIN/END PRIVATE KEY`
Depending on the version of openssl used to generate the management certificates
the private key may take one of two forms:

PKCS1 (Old, Currently Supported by us):

    -----BEGIN RSA PRIVATE KEY-----
    BASE64 ENCODED DATA
    -----END RSA PRIVATE KEY-----

PKCS8 (New, Support added in this commit):

    -----BEGIN PRIVATE KEY-----
    BASE64 ENCODED DATA
    -----END PRIVATE KEY-----

This commit adds support for the former. Prior to this commit using the above
format as part of a .pem would result in authentication failure as the key
would be undefined.