:art: Update myevents example
This commit updates myevents example project to Express 4.* and
latest version of Azure Node SDK module:
- update and bump up dependencies
- update implementation of server to Express 4
- update implementation of Azure SDK for Node to current version
and current routines
- small changes in implementation details for cleaner code and results
handling
- HTML markup update and rewrite to use HTML5 features, including
form validation features
- small code changes and fixes
- project README update to point to original version

The README has been updated as it contaisn a link to Channel 9 hosted
video post outlining older version of this project.
For this reason a git way to get detached version is added to readme.

Thanks!