Fixes for user agent header
1. ensure that this filter is executed last and other user-agent
filters, if any (xplat-cli has its own) are executed first. That way,
the right user agent header is set. So we call the other filters in the
chain first and apply our function when the callstack unwinds. By now,
if another filter had set the user agent header, we wont bother setting
it.

2. check if a particular bit of useragent info is already present before
inserting it.