URI encoding done before joining with comma
I was trying to use the `responseFilter` option in the `cognitiveServicesWebSearch` service and noticed I wasn't able to add multiple filters. After reading the code, I noticed the commas were part of the `encodeURIComponent` parameter, turning them into `%2C`. 

The API doesn't seem to decode them, so I instead mapped on each item of the array with `encodeURIComponent`, then join with comma.